+++ src/commands/projects/create.ts
@@ -1,5 +1,5 @@
 import type { ConfigInterface, DbConfigInterface } from '../../interfaces/project-create-interface';
-import type DumperV1 from '../../services/dumpers/dumper-v1';
+import type ForestExpress from '../../services/dumpers/forest-express';
 import type DatabaseAnalyzer from '../../services/schema/update/analyzer/database-analyzer';
 import type * as Config from '@oclif/config';
 
@@ -8,20 +8,20 @@ import AbstractProjectCreateCommand from '../../abstract-project-create-command'
 export default class CreateCommand extends AbstractProjectCreateCommand {
   private readonly databaseAnalyzer: DatabaseAnalyzer;
 
-  private readonly dumper: DumperV1;
+  private readonly dumper: ForestExpress;
 
   constructor(argv: string[], config: Config.IConfig, plan?) {
     super(argv, config, plan);
 
-    const { assertPresent, databaseAnalyzer, dumperV1 } = this.context;
+    const { assertPresent, databaseAnalyzer, forestExpressDumper } = this.context;
 
     assertPresent({
       databaseAnalyzer,
-      dumperV1,
+      forestExpressDumper,
     });
 
     this.databaseAnalyzer = databaseAnalyzer;
-    this.dumper = dumperV1;
+    this.dumper = forestExpressDumper;
   }
 
   async generateProject(config: ConfigInterface) {
@@ -30,22 +30,22 @@ export default class CreateCommand extends AbstractProjectCreateCommand {
   }
 
   private async analyzeDatabase(dbConfig: DbConfigInterface) {
-    let schema = {};
-
-    this.spinner.start({ text: 'Analyzing the database' });
     const connection = await this.database.connect(dbConfig);
 
     if (dbConfig.dbDialect === 'mongodb') {
       // the mongodb analyzer display a progress bar during the analysis
-      schema = await this.databaseAnalyzer.analyzeMongoDb(connection, dbConfig, true);
-    } else {
-      const schemaPromise = this.databaseAnalyzer.analyze(connection, dbConfig, true);
-      schema = await this.spinner.attachToPromise(schemaPromise);
+      this.logger.info('Analyzing the database');
+      const schema = await this.databaseAnalyzer.analyzeMongoDb(connection, dbConfig, true);
+      await this.database.disconnect(connection);
+      this.logger.success('Database is analyzed', { lineColor: 'green' });
+      return schema;
     }
 
-    await this.database.disconnect(connection);
+    this.spinner.start({ text: 'Analyzing the database' });
+    const schemaPromise = this.databaseAnalyzer.analyze(connection, dbConfig, true);
+    const schema = await this.spinner.attachToPromise(schemaPromise);
     this.logger.success('Database is analyzed', { lineColor: 'green' });
-
+    await this.database.disconnect(connection);
     return schema;
   }