+++ src/context/command-project-create-plan.js
@@ -5,11 +5,8 @@ module.exports = plan =>
     .addModule('mongodb', () => require('mongodb'))
     .addModule('Handlebars', () => require('handlebars'))
     .addUsingClass('database', () => require('../services/schema/update/database'))
-    .addUsingClass(
-      'agentNodejsDumper',
-      () => require('../services/dumpers/agent-nodejs-dumper').default,
-    )
-    .addUsingClass('dumperV1', () => require('../services/dumpers/dumper-v1'))
+    .addUsingClass('agentNodejsDumper', () => require('../services/dumpers/agent-nodejs').default)
+    .addUsingClass('forestExpressDumper', () => require('../services/dumpers/forest-express'))
     .addValue('GeneralPrompter', () => require('../services/prompter/general-prompter'))
     .addUsingClass('commandGenerateConfigGetter', () =>
       require('../services/projects/create/command-generate-config-getter'),