+++ src/context/utils-plan.js
@@ -6,6 +6,7 @@ module.exports = plan =>
     .addUsingClass('eventSender', () => require('../utils/event-sender'))
     .addUsingFunction('terminator', require('../utils/terminator'))
     .addValue('messages', require('../utils/messages'))
+    .addFunction('toValidPackageName', require('../utils/to-valid-package-name'))
+    .addUsingClass('strings', () => require('../utils/strings').default)
     .addFunction('buildDatabaseUrl', require('../utils/database-url').default)
-    .addFunction('isDatabaseLocal', require('../utils/database-url').isDatabaseLocal)
-    .addUsingClass('strings', () => require('../utils/strings').default);
+    .addFunction('isDatabaseLocal', require('../utils/database-url').isDatabaseLocal);