+++ src/services/dumpers/agent-nodejs.ts
@@ -1,11 +1,10 @@
 import type { ConfigInterface, DbConfigInterface } from '../../interfaces/project-create-interface';
 import type Strings from '../../utils/strings';
 
-import toValidPackageName from '../../utils/to-valid-package-name';
 import AbstractDumper from './abstract-dumper';
 
-export default class AgentNodeJsDumper extends AbstractDumper {
-  private env: { FOREST_SERVER_URL: string; FOREST_URL_IS_DEFAULT: string };
+export default class AgentNodeJs extends AbstractDumper {
+  private env: { FOREST_SERVER_URL: string; FOREST_URL_IS_DEFAULT: boolean };
 
   private readonly DEFAULT_PORT = 3310;
 
@@ -17,14 +16,23 @@ export default class AgentNodeJsDumper extends AbstractDumper {
 
   private readonly strings: Strings;
 
+  private readonly toValidPackageName: (string: string) => string;
+
   constructor(context) {
-    const { assertPresent, env, isLinuxOs, buildDatabaseUrl, isDatabaseLocal, strings } = context;
+    const {
+      assertPresent,
+      env,
+      isLinuxOs,
+      buildDatabaseUrl,
+      isDatabaseLocal,
+      strings,
+      toValidPackageName,
+    } = context;
 
     assertPresent({
       env,
       isLinuxOs,
       isDatabaseLocal,
-      buildDatabaseUrl,
       strings,
     });
 
@@ -35,6 +43,7 @@ export default class AgentNodeJsDumper extends AbstractDumper {
     this.buildDatabaseUrl = buildDatabaseUrl;
     this.isDatabaseLocal = isDatabaseLocal;
     this.strings = strings;
+    this.toValidPackageName = toValidPackageName;
   }
 
   // eslint-disable-next-line class-methods-use-this
@@ -69,7 +78,7 @@ export default class AgentNodeJsDumper extends AbstractDumper {
     }
 
     const pkg = {
-      name: toValidPackageName(applicationName),
+      name: this.toValidPackageName(applicationName),
       version: '0.0.1',
       private: true,
       main: 'index.js',
@@ -145,11 +154,15 @@ export default class AgentNodeJsDumper extends AbstractDumper {
   }
 
   private writeGitignore() {
-    this.writeFile('.gitignore', 'node_modules\n.env');
+    this.writeFile('.gitignore', 'node_modules\n.env\n');
+  }
+
+  private writeTypings() {
+    this.writeFile('typings.ts', '/* eslint-disable */\nexport type Schema = any;\n');
   }
 
   private writeDockerignore() {
-    this.writeFile('.dockerignore', 'node_modules\nnpm-debug.log\n.env');
+    this.writeFile('.dockerignore', 'node_modules\nnpm-debug.log\n.env\n');
   }
 
   private writeDockerfile() {
@@ -160,11 +173,10 @@ export default class AgentNodeJsDumper extends AbstractDumper {
     const databaseUrl = `\${${this.isLinuxOs ? 'DATABASE_URL' : 'DOCKER_DATABASE_URL'}}`;
     const forestServerUrl = this.env.FOREST_URL_IS_DEFAULT ? false : `\${FOREST_SERVER_URL}`;
 
-    let forestExtraHost: string | boolean = false;
+    let forestExtraHost: string = null;
     if (forestServerUrl) {
       try {
-        const parsedForestUrl = new URL(this.env.FOREST_SERVER_URL);
-        forestExtraHost = parsedForestUrl.hostname;
+        forestExtraHost = new URL(this.env.FOREST_SERVER_URL).hostname;
       } catch (error) {
         throw new Error(`Invalid value for FOREST_SERVER_URL: "${this.env.FOREST_SERVER_URL}"`);
       }
@@ -220,6 +232,7 @@ export default class AgentNodeJsDumper extends AbstractDumper {
       dumpConfig.forestEnvSecret,
       dumpConfig.forestAuthSecret,
     );
+    this.writeTypings();
     this.writeGitignore();
     this.writeDockerignore();
     this.writeDockerfile();