+++ src/utils/handlebars/helpers/or.js
@@ -1,3 +1,3 @@
 const Handlebars = require('handlebars');
 
-Handlebars.registerHelper('or', (value1, value2) => value1 || value2);
+Handlebars.registerHelper('or', (value1, value2) => Boolean(value1 || value2));